src/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-huggingface
langchain-text-splitters 
faiss-cpu
sentence-transformers
torch
transformers
accelerate
bitsandbytes
pydantic
python-dotenv